fix bug and some other annoyances
